Remove duplicate landscape module
